Create Number Guessing Game
Write a number guessing game. Player or the computer can play the game. The user will choose if he/she wants to play
or wants the computer to play. If the player chooses to play herself, the computer will generate a random integer between
1 and 100 and ask the user to make a guess. If user's guess is higher than the number, it will print something like 'go lower'
and if it is low, print something in the lines of 'go higher'. This will continue until the user guesses the number, in
which case the computer will print 'Correct. It took you XXXX tries' and ask if the user wants to go again. If yes, it will
start over and ask who will play the game again. If no, the program will terminate. You should check if the number entered by the user is
indeed between 1 and 100, and notify the user if it is not.

If the user wants the computer to play the game, your script will ask the user to provide a number between 1 and 100. Check
if the number the user entered is between these values. If not, warn the user and ask again. If it is, the program should
guess the number until it finds it. It is up to you which algorithm your program will use but try to find a reasonably efficient
algorithm (i.e. do not use 'for i in range(1, 101): if i == number.......). Print every number it tries and if it is lower or
higher than your number .When the computer makes the right guess, print how many tries it took the computer to find the
number on the screen.