Create DNA Sequence Validity Check
Take a DNA sequence from the user and print if the sequence is a valid sequence or not on the screen in the following format. Provide the user with the invalid character and its position for user's convenience. If there are several invalid bases, use the first one.

---Ex for valid sequence:

Please type a DNA sequence: TGTGAGTATGCTAGC
This is a valid DNA sequence

---Ex for invalid sequence:

Please type a DNA sequence: TXGTCTXGACGGAC
X (base 2) is not a valid base. This is not a valid DNA sequence